Moved stuff from asb and "helpfull"functions to new super class
