+++ spinn_front_end_common/utilities/helpful_functions.py
@@ -1,8 +1,6 @@
 import os
 import logging
 import struct
-import datetime
-import shutil
 from spinn_utilities.log import FormatAdapter
 from spinn_machine import CoreSubsets
 from spinnman.model.enums import CPUState
@@ -11,10 +9,6 @@ from spinn_front_end_common.utilities.exceptions import ConfigurationException
 from spinn_front_end_common.utilities.utility_objs import ExecutableType
 
 logger = FormatAdapter(logging.getLogger(__name__))
-APP_DIRNAME = 'application_generated_data_files'
-FINISHED_FILENAME = "finished"
-REPORTS_DIRNAME = "reports"
-TIMESTAMP_FILENAME = "time_stamp"
 _ONE_WORD = struct.Struct("<I")
 
 
@@ -76,160 +70,6 @@ def locate_memory_region_for_placement(placement, region, transceiver):
     region_address = transceiver.read_memory(
         placement.x, placement.y, region_offset, 4)
     return _ONE_WORD.unpack_from(region_address)[0]
-
-
-def child_folder(parent, child_name):
-    child = os.path.join(parent, child_name)
-    if not os.path.exists(child):
-        os.makedirs(child)
-    return child
-
-
-def set_up_output_application_data_specifics(
-        where_to_write_application_data_files,
-        max_application_binaries_kept, n_calls_to_run,
-        this_run_time_string):
-    """
-    :param where_to_write_application_data_files:\
-        the location where all app data is by default written to, or DEFAULT
-    :type where_to_write_application_data_files: str
-    :param max_application_binaries_kept:\
-        The max number of report folders to keep active at any one time
-    :type max_application_binaries_kept: int
-    :param n_calls_to_run: the counter of how many times run has been called.
-    :type n_calls_to_run: int
-    :param this_run_time_string: the time stamp string for this run
-    :type this_run_time_string: str
-    :return: the run folder for this simulation to hold app data
-    """
-    if where_to_write_application_data_files == "DEFAULT":
-        where_to_write_application_data_files = os.getcwd()
-    application_generated_data_file_folder = child_folder(
-        where_to_write_application_data_files, APP_DIRNAME)
-    # add time stamped folder for this run
-    this_run_time_folder = child_folder(
-        application_generated_data_file_folder, this_run_time_string)
-
-    # remove folders that are old and above the limit
-    _remove_excess_folders(
-        max_application_binaries_kept,
-        application_generated_data_file_folder)
-
-    # store timestamp in latest/time_stamp
-    time_of_run_file_name = os.path.join(
-        this_run_time_folder, TIMESTAMP_FILENAME)
-    with open(time_of_run_file_name, "w") as f:
-        f.writelines(str(this_run_time_string))
-
-    # create sub folder within reports for sub runs (where changes need to be
-    # recorded)
-    this_run_time_sub_folder = child_folder(
-        this_run_time_folder, "run_{}".format(n_calls_to_run))
-
-    return this_run_time_sub_folder, this_run_time_folder
-
-
-def set_up_report_specifics(
-        default_report_file_path, max_reports_kept, n_calls_to_run,
-        this_run_time_string=None):
-    """
-    :param default_report_file_path: The location where all reports reside
-    :type default_report_file_path: str
-    :param max_reports_kept:\
-        The max number of report folders to keep active at any one time
-    :type max_reports_kept: int
-    :param n_calls_to_run: the counter of how many times run has been called.
-    :type n_calls_to_run: int
-    :param this_run_time_string: holder for the timestamp for future runs
-    :type this_run_time_string: str
-    :return: The folder for this run, the time_stamp
-    """
-
-    # determine common report folder
-    config_param = default_report_file_path
-    if config_param == "DEFAULT":
-        directory = os.getcwd()
-
-        # global reports folder
-        report_default_directory = child_folder(directory, REPORTS_DIRNAME)
-    elif config_param == "REPORTS":
-        report_default_directory = REPORTS_DIRNAME
-        if not os.path.exists(report_default_directory):
-            os.makedirs(report_default_directory)
-    else:
-        report_default_directory = child_folder(config_param, REPORTS_DIRNAME)
-
-    # clear and clean out folders considered not useful anymore
-    if os.listdir(report_default_directory):
-        _remove_excess_folders(max_reports_kept, report_default_directory)
-
-    # determine the time slot for later
-    if this_run_time_string is None:
-        now = datetime.datetime.now()
-        this_run_time_string = (
-            "{:04}-{:02}-{:02}-{:02}-{:02}-{:02}-{:02}".format(
-                now.year, now.month, now.day,
-                now.hour, now.minute, now.second, now.microsecond))
-
-    # handle timing app folder and cleaning of report folder from last run
-    app_folder_name = child_folder(
-        report_default_directory, this_run_time_string)
-
-    # create sub folder within reports for sub runs (where changes need to be
-    # recorded)
-    app_sub_folder_name = child_folder(
-        app_folder_name, "run_{}".format(n_calls_to_run))
-
-    # store timestamp in latest/time_stamp for provenance reasons
-    time_of_run_file_name = os.path.join(app_folder_name, TIMESTAMP_FILENAME)
-    with open(time_of_run_file_name, "w") as f:
-        f.writelines(this_run_time_string)
-    return app_sub_folder_name, app_folder_name, this_run_time_string
-
-
-def write_finished_file(app_data_runtime_folder, report_default_directory):
-    # write a finished file that allows file removal to only remove folders
-    # that are finished
-    app_file_name = os.path.join(app_data_runtime_folder, FINISHED_FILENAME)
-    with open(app_file_name, "w") as f:
-        f.writelines("finished")
-
-    app_file_name = os.path.join(report_default_directory, FINISHED_FILENAME)
-    with open(app_file_name, "w") as f:
-        f.writelines("finished")
-
-
-def _remove_excess_folders(max_to_keep, starting_directory):
-    files_in_report_folder = os.listdir(starting_directory)
-
-    # while there's more than the valid max, remove the oldest one
-    if len(files_in_report_folder) > max_to_keep:
-
-        # sort files into time frame
-        files_in_report_folder.sort(
-            key=lambda temp_file:
-            os.path.getmtime(os.path.join(starting_directory, temp_file)))
-
-        # remove only the number of files required, and only if they have
-        # the finished flag file created
-        num_files_to_remove = len(files_in_report_folder) - max_to_keep
-        files_removed = 0
-        files_not_closed = 0
-        for current_oldest_file in files_in_report_folder:
-            finished_flag = os.path.join(os.path.join(
-                starting_directory, current_oldest_file), FINISHED_FILENAME)
-            if os.path.exists(finished_flag):
-                shutil.rmtree(
-                    os.path.join(starting_directory, current_oldest_file),
-                    ignore_errors=True)
-                files_removed += 1
-            else:
-                files_not_closed += 1
-            if files_removed + files_not_closed >= num_files_to_remove:
-                break
-        if files_not_closed > max_to_keep // 4:
-            logger.warning("{} has {} old reports that have not been closed",
-                           starting_directory, files_not_closed)
 
 
 def convert_string_into_chip_and_core_subset(cores):
